Finished TossCoin function
The program has a head button and tail button at the start. If you click the head button, the user chooses heads & if you click the blue button, the user chooses tail. The program makes the attack button appear if the user chooses heads/tails, and the coin toss results in heads/tails. If the user chooses heads/tails, and the coin toss results in tails/heads, the program makes both the attack button and defend button appear